chore: Move logic to OperationParamUtils [TECH-1750]
